TASK-67 [Rename]: column,card api호출함수 파일 세분화
